Move connect and adapters/converters to sqlite.database
