+++ qcodes/dataset/sqlite_base.py
@@ -1,13 +1,10 @@
-import sys
 import logging
 import sqlite3
 import time
-import io
 import warnings
 from typing import (Any, List, Optional, Tuple, Union, Dict, cast, Callable,
                     Sequence)
 
-from numpy import ndarray
 import numpy as np
 
 import qcodes as qc
@@ -19,6 +16,7 @@ from qcodes.dataset.guids import generate_guid, parse_guid
 from qcodes.dataset.sqlite.connection import ConnectionPlus, atomic, \
     transaction, atomic_transaction
 from qcodes.dataset.sqlite.connection import make_connection_plus_from
+from qcodes.dataset.sqlite.database import connect
 from qcodes.dataset.sqlite.db_upgrades import _latest_available_version, \
     perform_db_upgrade, upgrader
 from qcodes.dataset.sqlite.db_upgrades.version import get_user_version
@@ -33,7 +31,6 @@ from qcodes.dataset.sqlite.query_helpers import insert_many_values, VALUE, \
 from qcodes.dataset.sqlite.db_upgrades import perform_db_upgrade_0_to_1, \
     perform_db_upgrade_1_to_2, perform_db_upgrade_2_to_3, \
     perform_db_upgrade_3_to_4, perform_db_upgrade_4_to_5
-from qcodes.utils.types import complex_types, complex_type_union
 
 
 log = logging.getLogger(__name__)
@@ -55,156 +52,6 @@ def sql_placeholder_string(n: int) -> str:
     Example: sql_placeholder_string(5) returns '(?,?,?,?,?)'
     """
     return '(' + ','.join('?'*n) + ')'
-
-
-# utility function to allow sqlite/numpy type
-def _adapt_array(arr: ndarray) -> sqlite3.Binary:
-    """
-    See this:
-    https://stackoverflow.com/questions/3425320/sqlite3-programmingerror-you-must-not-use-8-bit-bytestrings-unless-you-use-a-te
-    """
-    out = io.BytesIO()
-    np.save(out, arr)
-    out.seek(0)
-    return sqlite3.Binary(out.read())
-
-
-def _convert_array(text: bytes) -> ndarray:
-    out = io.BytesIO(text)
-    out.seek(0)
-    return np.load(out)
-
-
-def _convert_complex(text: bytes) -> complex_type_union:
-    out = io.BytesIO(text)
-    out.seek(0)
-    return np.load(out)[0]
-
-
-this_session_default_encoding = sys.getdefaultencoding()
-
-
-def _convert_numeric(value: bytes) -> Union[float, int, str]:
-    """
-    This is a converter for sqlite3 'numeric' type class.
-
-    This converter is capable of deducting whether a number is a float or an
-    int.
-
-    Note sqlite3 allows to save data to columns even if their type is not
-    compatible with the table type class (for example, it is possible to save
-    integers into 'text' columns). Due to this fact, and for the reasons of
-    flexibility, the numeric converter is also made capable of handling
-    strings. An obvious exception to this is 'nan' (case insensitive) which
-    gets converted to `np.nan`. Another exception to this is 'inf', which
-    gets converted to 'np.inf'.
-    """
-    try:
-        # First, try to convert bytes to float
-        numeric = float(value)
-    except ValueError as e:
-        # If an exception has been raised, we first need to find out
-        # if the reason was the conversion to float, and, if so, we are sure
-        # that we need to return a string
-        if "could not convert string to float" in str(e):
-            return str(value, encoding=this_session_default_encoding)
-        else:
-            # otherwise, the exception is forwarded up the stack
-            raise e
-
-    # If that worked, e.g. did not raise an exception, then we check if the
-    # outcome is 'nan'
-    if np.isnan(numeric):
-        return numeric
-
-    # Then we check if the outcome is 'inf', includes +inf and -inf
-    if np.isinf(numeric):
-        return numeric
-
-    # If it is not 'nan' and not 'inf', then we need to see if the value is
-    # really an integer or with floating point digits
-    numeric_int = int(numeric)
-    if numeric != numeric_int:
-        return numeric
-    else:
-        return numeric_int
-
-
-def _adapt_float(fl: float) -> Union[float, str]:
-    if np.isnan(fl):
-        return "nan"
-    return float(fl)
-
-
-def _adapt_complex(value: complex_type_union) -> sqlite3.Binary:
-    out = io.BytesIO()
-    np.save(out, np.array([value]))
-    out.seek(0)
-    return sqlite3.Binary(out.read())
-
-
-def connect(name: str, debug: bool = False,
-            version: int = -1) -> ConnectionPlus:
-    """
-    Connect or create  database. If debug the queries will be echoed back.
-    This function takes care of registering the numpy/sqlite type
-    converters that we need.
-
-    Args:
-        name: name or path to the sqlite file
-        debug: whether or not to turn on tracing
-        version: which version to create. We count from 0. -1 means 'latest'.
-            Should always be left at -1 except when testing.
-
-    Returns:
-        conn: connection object to the database (note, it is
-            `ConnectionPlus`, not `sqlite3.Connection`
-
-    """
-    # register numpy->binary(TEXT) adapter
-    # the typing here is ignored due to what we think is a flaw in typeshed
-    # see https://github.com/python/typeshed/issues/2429
-    sqlite3.register_adapter(np.ndarray, _adapt_array)  # type: ignore
-    # register binary(TEXT) -> numpy converter
-    # for some reasons mypy complains about this
-    sqlite3.register_converter("array", _convert_array)
-
-    sqlite3_conn = sqlite3.connect(name, detect_types=sqlite3.PARSE_DECLTYPES)
-    conn = ConnectionPlus(sqlite3_conn)
-
-    latest_supported_version = _latest_available_version()
-    db_version = get_user_version(conn)
-
-    if db_version > latest_supported_version:
-        raise RuntimeError(f"Database {name} is version {db_version} but this "
-                           f"version of QCoDeS supports up to "
-                           f"version {latest_supported_version}")
-
-    # sqlite3 options
-    conn.row_factory = sqlite3.Row
-
-    # Make sure numpy ints and floats types are inserted properly
-    for numpy_int in [
-        np.int, np.int8, np.int16, np.int32, np.int64,
-        np.uint, np.uint8, np.uint16, np.uint32, np.uint64
-    ]:
-        sqlite3.register_adapter(numpy_int, int)
-
-    sqlite3.register_converter("numeric", _convert_numeric)
-
-    for numpy_float in [np.float, np.float16, np.float32, np.float64]:
-        sqlite3.register_adapter(numpy_float, _adapt_float)
-
-    for complex_type in complex_types:
-        sqlite3.register_adapter(complex_type, _adapt_complex)  # type: ignore
-    sqlite3.register_converter("complex", _convert_complex)
-
-    if debug:
-        conn.set_trace_callback(print)
-
-    init_db(conn)
-    perform_db_upgrade(conn, version=version)
-    return conn
 
 
 def get_db_version_and_newest_available_version(path_to_db: str) -> Tuple[int,
